requirements.txt
streamlit>=1.32
numpy>=1.26
pandas>=2.2
Pillow>=10.0
plotly>=5.20
